Added files for week 2 & 3 for modeling and data
These two weeks include how to focus on data-centric development
as well as the importance of correct labeling for small data sets.

Week 2 goes deep into keeping the model the same but applying either
data-augmentation techniques or showing how class imbalance affects
the model.

Week 3 focused on the importance of correct labeling and guidelines
and showed us that even if the ground truth is not given, we can
make approximations using performance lower bounds (random) and
little by little change the quality of our data to imporve model
performance. Also, a very in-house way to do experiment tracking/
VCS for ML.

There are some reallyyyyy cool stuff in the notes about tips for
experiment tracking and error analysis. DO NOT forget to use this
in a later project.